+++ extras/display_code/dollar.bash
@@ -1,0 +1,5 @@
+# A comment field
+$ ls
+stuff/ things/
+(VENV) $ ls
+stuff/ things/